Add InitialSyncCount and refactor services/UI
Introduce InitialSyncCount (default 500, 0 = fetch all) to ConfigModel and wire parsing/serialization in ConfigService. Change IImapService.GetMessagesSinceAsync to accept an initialCount and update ImapService, SyncService, MainViewModel and test stubs accordingly. Add AddressParser and MimeMessageBuilder to centralize MIME/address construction and replace duplicated logic in SmtpService and ImapService. Add AccountEditorViewModel to consolidate account-editing fields/commands and refactor AccountManagerViewModel and AddAccountViewModel to inherit from it. Add TreeViewFocusHelper to centralize TreeView selection/type-ahead utilities and update MainWindow/FolderPickerWindow to use it. These changes reduce duplication and make initial-folder sync behavior configurable (note: IImapService signature changed).
