+++ QuickMail/Services/ImapService.cs
@@ -220,7 +220,7 @@ public class ImapService : IImapService
     }
 
     public Task<List<MailMessageSummary>> GetMessagesSinceAsync(
-        Guid accountId, string folderName, uint sinceUid, CancellationToken ct = default)
+        Guid accountId, string folderName, uint sinceUid, int initialCount, CancellationToken ct = default)
         => ExecuteWithRetryAsync(accountId, ct, async client =>
         {
             var folder = await client.GetFolderAsync(folderName, ct);
@@ -236,7 +236,8 @@ public class ImapService : IImapService
                 if (sinceUid == 0)
                 {
                     if (folder.Count == 0) return new List<MailMessageSummary>();
-                    var startIndex = Math.Max(0, folder.Count - 500);
+                    var count = initialCount > 0 ? initialCount : folder.Count;
+                    var startIndex = Math.Max(0, folder.Count - count);
                     summaries = await folder.FetchAsync(startIndex, -1, items, ct);
                 }
                 else
@@ -372,54 +373,7 @@ public class ImapService : IImapService
         var draftsFolder = FindSpecialFolder(client, SpecialFolder.Drafts)
             ?? throw new InvalidOperationException("No Drafts folder found for this account.");
 
-        // Build the MIME message from compose fields
-        var msg = new MimeMessage();
-        msg.From.Add(new MailboxAddress(account.SenderDisplayName, account.Username));
-
-        static void AddAddresses(InternetAddressList list, string raw)
-        {
-            if (string.IsNullOrWhiteSpace(raw)) return;
-            foreach (var part in raw.Split(';', ',').Select(a => a.Trim()).Where(a => a.Length > 0))
-            {
-                try { list.Add(InternetAddress.Parse(part)); }
-                catch { /* skip unparseable address */ }
-            }
-        }
-
-        AddAddresses(msg.To,  draft.To);
-        AddAddresses(msg.Cc,  draft.Cc);
-        AddAddresses(msg.Bcc, draft.Bcc);
-
-        msg.Subject = draft.Subject;
-
-        if (!string.IsNullOrEmpty(draft.InReplyToMessageId))
-            msg.InReplyTo = draft.InReplyToMessageId;
-
-        var loadedAttachments = draft.Attachments.Where(a => a.IsLoaded).ToList();
-        if (loadedAttachments.Count > 0)
-        {
-            var multipart = new Multipart("mixed");
-            multipart.Add(new TextPart("plain") { Text = draft.Body });
-            foreach (var att in loadedAttachments)
-            {
-                var slash = att.ContentType.IndexOf('/');
-                var mediaType    = slash >= 0 ? att.ContentType[..slash] : "application";
-                var mediaSubtype = slash >= 0 ? att.ContentType[(slash + 1)..] : "octet-stream";
-                var mimePart = new MimePart(mediaType, mediaSubtype)
-                {
-                    Content                  = new MimeContent(new MemoryStream(att.Content!)),
-                    ContentDisposition       = new ContentDisposition(ContentDisposition.Attachment),
-                    ContentTransferEncoding  = ContentEncoding.Base64,
-                    FileName                 = att.FileName,
-                };
-                multipart.Add(mimePart);
-            }
-            msg.Body = multipart;
-        }
-        else
-        {
-            msg.Body = new TextPart("plain") { Text = draft.Body };
-        }
+        var msg = MimeMessageBuilder.Build(draft, account);
 
         // Delete the old draft revision before appending the new one
         if (replaceUid.HasValue)
