+++ QuickMail/ViewModels/AccountEditorViewModel.cs
@@ -1,0 +1,126 @@
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+using CommunityToolkit.Mvvm.ComponentModel;
+using CommunityToolkit.Mvvm.Input;
+using QuickMail.Models;
+using QuickMail.Services;
+
+namespace QuickMail.ViewModels;
+
+/// <summary>
+/// Base class for account-editor ViewModels, containing the shared form fields,
+/// auth-type helpers, and OAuth / connection-test commands.
+/// </summary>
+public abstract partial class AccountEditorViewModel : ObservableObject
+{
+    protected readonly IImapService ImapService;
+    protected readonly IOAuthService OAuthService;
+
+    [ObservableProperty] private string _accountName = string.Empty;
+    [ObservableProperty] private string _displayName = string.Empty;
+    [ObservableProperty] private string _username = string.Empty;
+    [ObservableProperty] private string _password = string.Empty;
+    [ObservableProperty] private string _imapHost = string.Empty;
+    [ObservableProperty] private int    _imapPort = 993;
+    [ObservableProperty] private bool   _imapUseSsl = true;
+    [ObservableProperty] private bool   _imapAcceptInvalidCert = false;
+    [ObservableProperty] private string _smtpHost = string.Empty;
+    [ObservableProperty] private int    _smtpPort = 587;
+    [ObservableProperty] private bool   _smtpUseSsl = false;
+    [ObservableProperty] private bool   _smtpAcceptInvalidCert = false;
+
+    [ObservableProperty]
+    [NotifyPropertyChangedFor(nameof(IsPasswordAuth))]
+    [NotifyPropertyChangedFor(nameof(IsOAuth2))]
+    [NotifyPropertyChangedFor(nameof(AuthTypeIndex))]
+    private AuthType _authType = AuthType.Password;
+
+    [ObservableProperty] private string _statusText = string.Empty;
+    [ObservableProperty] private bool   _isBusy = false;
+
+    public bool IsPasswordAuth => AuthType == AuthType.Password;
+    public bool IsOAuth2       => AuthType == AuthType.OAuth2Microsoft;
+
+    public int AuthTypeIndex
+    {
+        get => AuthType == AuthType.Password ? 0 : 1;
+        set => AuthType = value == 0 ? AuthType.Password : AuthType.OAuth2Microsoft;
+    }
+
+    protected AccountEditorViewModel(IImapService imap, IOAuthService oauth)
+    {
+        ImapService = imap;
+        OAuthService = oauth;
+    }
+
+    /// <summary>
+    /// Called when <see cref="AuthType"/> changes. Override in derived classes
+    /// to react (e.g. auto-fill server settings for OAuth).
+    /// </summary>
+    protected virtual void OnAuthTypeChangedInternal(AuthType value) { }
+
+    partial void OnAuthTypeChanged(AuthType value) => OnAuthTypeChangedInternal(value);
+
+    [RelayCommand]
+    private async Task SignInMicrosoftAsync()
+    {
+        IsBusy = true;
+        StatusText = "Opening browser for Microsoft sign-in…";
+        try
+        {
+            using var cts = new CancellationTokenSource(TimeSpan.FromMinutes(3));
+            var tempAccount = new AccountModel { Username = Username, AuthType = AuthType.OAuth2Microsoft };
+            var result = await OAuthService.SignInInteractiveAsync(tempAccount, cts.Token);
+            Username = result.Username;
+            StatusText = $"Signed in as {result.Username}";
+        }
+        catch (Exception ex)
+        {
+            StatusText = $"Sign-in failed: {ex.Message}";
+        }
+        finally
+        {
+            IsBusy = false;
+        }
+    }
+
+    [RelayCommand]
+    private async Task TestConnectionAsync()
+    {
+        if (string.IsNullOrWhiteSpace(ImapHost) || string.IsNullOrWhiteSpace(Username))
+        {
+            StatusText = "Fill in IMAP host and username first.";
+            return;
+        }
+
+        IsBusy = true;
+        StatusText = "Testing connection…";
+        try
+        {
+            var testAccount = new AccountModel
+            {
+                Id = Guid.NewGuid(),
+                Username = Username,
+                AuthType = AuthType,
+                ImapHost = ImapHost,
+                ImapPort = ImapPort,
+                ImapUseSsl = ImapUseSsl,
+                ImapAcceptInvalidCert = ImapAcceptInvalidCert
+            };
+            using var cts = new CancellationTokenSource(TimeSpan.FromSeconds(30));
+            var pwd = IsPasswordAuth ? Password : null;
+            await ImapService.ConnectAsync(testAccount, pwd, cts.Token);
+            await ImapService.DisconnectAsync(testAccount.Id, cts.Token);
+            StatusText = "Connection successful!";
+        }
+        catch (Exception ex)
+        {
+            StatusText = $"Connection failed: {ex.Message}";
+        }
+        finally
+        {
+            IsBusy = false;
+        }
+    }
+}
